Update 0.3.0
First working update for project manager and engine manager. Works for 4.3, need to add older version support

+++ assets/scripts/components/engine/engine.gd
@@ -1,7 +1,6 @@
 extends Node
 
 const DEFAULT_ICON: CompressedTexture2D = preload("res://icon.svg")
-const MISSING_ICON: CompressedTexture2D = preload("res://assets/icons/missing_icon.svg")
 
 @onready var icon: TextureRect = $EngineContainer/Icon
 @onready var title_label: Label = $EngineContainer/Info/Title
@@ -15,7 +14,7 @@ var master: Control = null
 var this_engine: int = -1
 
 
-func setup(new_master: Control, engine: int, new_title: String, source: String, installed: bool) -> void:
+func setup(new_master: Control, engine: int, new_title: String, source: String, installed: bool, default: bool) -> void:
 	if not is_node_ready():
 		await ready
 	master = new_master
@@ -25,15 +24,36 @@ func setup(new_master: Control, engine: int, new_title: String, source: String, 
 	download_button.set_visible(!installed)
 	uninstall_button.set_visible(installed)
 	default_button.set_visible(installed)
+	default_button.set_pressed_no_signal(default)
 
 
 func button_pressed(button: String) -> void:
 	match button:
 		"download":
-			var complete = EngineManager.install_version(this_engine)
-			if not complete:
-				NotificationManager.show_prompt("Failed To Install, Verify Install Path.", ["OK"], self, "")
+			NotificationManager.show_prompt("Download And Install Version?", ["No", "Yes"], self, "_on_install")
 		"uninstall":
-			EngineManager.uninstall_version(this_engine)
-		"default":
-			pass
+			NotificationManager.show_prompt("Uninstall Version?", ["No", "Yes"], self, "_on_uninstall")
+
+
+func _on_default_button_toggled(toggled_on: bool) -> void:
+	if toggled_on:
+		ConfigManager.set_config_data("settings", "default_engine", title_label.get_text())
+		master.reset_default_engine(self)
+	else:
+		default_button.set_pressed_no_signal(true)
+
+
+func _on_install(option: String) -> void:
+	if option == "Yes":
+		NotificationManager.notify("Downloading And Installing...", 10.0, true)
+		var complete = EngineManager.install_version(this_engine)
+		if not complete:
+			NotificationManager.show_prompt("Failed To Install, Verify Install Path.", ["OK"], self, "")
+
+
+func _on_uninstall(option: String) -> void:
+	if option == "Yes":
+		NotificationManager.notify("Uninstalling...", 10.0, true)
+		var complete = EngineManager.uninstall_version(this_engine)
+		if not complete:
+			NotificationManager.show_prompt("Failed To Uninstall, Verify Install Path.", ["OK"], self, "")
